requirements.txt
django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
django-environ
django-storages[boto3]
drf-spectacular
django-simple-history
celery
django-role-permissions
pytest
factory-boy
mock
django-cors-headers
python-dateutil
pillow
pytz
